Fix user star ratings and long stuff

+++ totally-legit-horoscopes-api/DataAccess/Interfaces/IGenericRepository.cs
@@ -5,10 +5,10 @@ namespace totally_legit_horoscopes_api.DataAccess
 {
     public interface IGenericRepository<T> where T : class
     {
-        Task<T> Get(int id);
+        Task<T> Get(long id);
         Task<IEnumerable<T>> GetAll();
         Task<bool> Add(T entity);
-        Task<bool> Delete(int id);
+        Task<bool> Delete(long id);
         Task<bool> Update(T entity);
         void Save();
